Start upload background task together with writing task to prevent a logic race condition.
If the upload task is started by the writer task enqueueing the first completed file (as before), a FinishAsync call can observe a null uploadTask and thus not await the uploader task.
When the upload task is then started, it can block or delete files that are supposed to be finished.