Changed way of disabling log collector client from a null client reference to using an IsEnabled property on the client object to allow sensible default for the client parameter.
The field in SGLAnalytics for the client is now not-nullable and thus always contains a client object.
Using an IsEnabled property on that object allows the null-value for the constructor parameter to mean the sensible default of using the real REST implementation for the client instead of defaulting to disabled uploads.