Fixed navigation properties on entity classes.
They can't be bound using constructor arguments.
To preserve their non-optional character, they can also not be made nullable.
Therefore the constructor produces a warning about non-nullable properties that are not initialized to something non-null.
Following the documentation (https://docs.microsoft.com/en-us/ef/core/miscellaneous/nullable-reference-types#non-nullable-properties-and-initialization), they are initialized with the null-forgiving operator.
That null value is however only seen in case of a programming bug, where the property is not fetched by a query, but used anyway.