Removed unnecessary Dispose methods for DbApplicationRepository and DbLogMetadataRepository.
The DbContext doesn't need to (and should not) be disposed by the repository object.
In the real usage scenario it is disposed by the DI container.
In the test scenario it should be disposed by the test code instead of the test code disposing it through the repository.
Theoretically, one might want to use two different repositories with differing lifetime on the same DbContext, which would be broken by each repository object disposing the context.