Added fallback value getter in NamedPlaceholderFormatterFactory<T>.
This getter is called if the placeholder name is not registered.
The name is then passed to the getter and consuming code can decide where to look for the value for that name.
If the fallback getter is not set, the old behavior of throwing an exception during format string parsing is used.