Bugfix: Ensure the correct status code is returned when a log file upload exceeds the file size limit.
Although normally (if not caught) the BadHttpRequestException would bubble up out of the controller code and Kestrel would return a 413 as expected,
this seemingly doesn't work for some reason if the exception is caught and rethrown, because it seems to go in the regular uncaught exception -> 500 error path.
To fix this, explicitly return a 413 error code result instead of rethrowing the exception.