Added example recipient certificates for local development and testing.
Note:
This commit also adds the associated private keys for these TESTING certificates.
As they are intended for local testing on development machines, the associated private keys are also present in the repository to allow checking of encrypted test data.
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
THESE CERTIFICATES MUST ABSOLUTELY NOT BE USED IN PRODUCTION UNDER ANY CIRCUMSTANCES!
!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!
Those example certificates are provided as a convenience to avoid having to setup the trust chain locally for each test run.
For this reason, the associated signer certificate is also included.