Added UseAuthentication() to the user registration service request pipeline.
The service now does not only perform the user registration and login (which both can't themselves be authenticated requests).
It now also handles exporter requests for user registration data, which are authenticated using a JWT bearer token issued by ExporterKeyAuthController.
We therefore now need checking of these tokens and injection of the contained claims in the pipeline for authentication to work correctly.