Fixed default value handling for log content encoding.
Previously, LogContentEncoding.Plain was mapped to 0, combined with .HasDefaultValue(LogContentEncoding.GZipCompressed) configured in the model,
this caused a serious bug, where Plain was replaced by with GZipCompressed by EF Core when a record is inserted.
This happened because Plain was the default value in code, when a LogContentEncoding was not initialized
and thus EF Core could not differentiate between an uninitialized LogContentEncoding and one explicitly set to Plain.
Therefor old records that were intended as Plain (0) are in the database as GZipCompressed (1) and Plain (0) is not present in existing records.
To fix this for future records, we change Plain to map to 2 instead, making it non-default.
Existing records can't automatically be fixed, but luckily
 - Plain records are quite rare, as the code used GZipCompressed by-default.
 - we are still in staging only, so the impact is low


(cherry picked from commit f2b06785038c7523e72e42dc291dc7a804e1c22a)