Changed signatures of GetDecryptedLogFilesAsync and GetDecryptedUserRegistrationsAsync:
An outer async method evaluates the query Action and asynchronously obtains the current per-app state.
It then returns an IAsyncEnumerable for the inner method that is invoked when the iteration starts, with the per-app state and configured query parameters passed-in.
Thus, the query params are evaluated in the caller's context and per-app state is captured immediately.
This should protect from an app switch between calling the outer function and starting the iteration.